ft_strnstr

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: lamorim <lamorim@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/02 14:37:18 by lamorim           #+#    #+#             */
-/*   Updated: 2021/08/02 14:37:22 by lamorim          ###   ########.fr       */
+/*   Updated: 2021/08/03 18:26:26 by lamorim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	while ((s1[i] != '\0' || s2[i] != '\0') && i < n)
 	{
 		if (s1[i] != s2[i])
-			return ((int)(size_t)s1[i] - (int)(size_t)s2[i]);
+			return ((int)s1[i] - (int)s2[i]);
 		i++;
 	}
 	return (0);
